+++ rotkehlchen/tasks/utils.py
@@ -26,6 +26,7 @@ def should_run_periodic_task(
             'last_evm_accounts_detect_ts',
             'last_spam_assets_detect_key',
             'last_augmented_spam_assets_detect_key',
+            'last_owned_assets_update',
         ],
         refresh_period: int,
 ) -> bool: